requirements.txt
tushare>=1.2.89
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
plotly>=5.3.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0 